pyproject.toml
[project]
name = "adk-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.8.0",
    "litellm>=1.74.9",
    "pydantic>=2.11.7",
]
